requirements.txt
# dependencies
asyncio==3.4.3
beautifulsoup4==4.12.2
colorama==0.4.6
duckduckgo_search==3.8.4
md2pdf==1.0.1
openai~=0.27.8
playwright==1.36.0
python-dotenv~=1.0.0
pyyaml==6.0.1
selenium==4.11.2
webdriver-manager==4.0.0
Flask==2.3.2
uvicorn==0.23.2
pydantic==1.10.12
fastapi==0.101.1
python-multipart==0.0.6
markdown
langchain==0.0.263
tenacity==8.2.3
